move DA code to vaemixin
